pyproject.toml
[project]
name = "fast"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.136.1",
    "jinja2>=3.1.6",
    "langchain>=1.2.15",
    "langchain-deepseek>=1.0.1",
    "langchain-openai>=1.2.0",
    "langchain-tavily>=0.2.18",
    "langgraph>=1.1.9",
    "langgraph-checkpoint-sqlite>=3.0.3",
    "numpy>=2.4.4",
    "python-dotenv>=1.2.2",
    "scipy>=1.17.1",
    "scipy-stubs>=1.17.1.4",
    "uvicorn>=0.46.0",
]

[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true
